api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
affine==2.3.0
    # via
    #   -r requirements.in
    #   rasterio
amqp==5.0.6
    # via
    #   -r requirements.in
    #   kombu
appdirs==1.4.4
    # via -r requirements.in
argon2-cffi==25.1.0
    # via minio
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.8.1
    # via django
attrs==21.2.0
    # via
    #   -r requirements.in
    #   fiona
    #   jsonschema
    #   rasterio
babel==2.17.0
    # via pygeoapi
beautifulsoup4==4.13.4
    # via pyoscar
billiard==3.6.4.0
    # via celery
blinker==1.9.0
    # via flask
cartopy==0.21.1
    # via -r requirements.in
celery==5.2.1
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.6.15
    # via
    #   fiona
    #   minio
    #   pyproj
    #   rasterio
    #   requests
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   fiona
    #   flask
    #   pygeoapi
    #   pyoscar
    #   rasterio
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   fiona
    #   rasterio
click-repl==0.3.0
    # via celery
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
cronex==0.1.3.1
    # via -r requirements.in
croniter==1.3.5
    # via -r requirements.in
cryptography==45.0.4
    # via -r requirements.in
cycler==0.12.1
    # via matplotlib
dateparser==1.2.0
    # via pygeofilter
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20241021
    # via django-import-export
django==3.2.13
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-ckeditor
    #   django-cors-headers
    #   django-import-export
    #   django-js-asset
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-celery-beat==2.2.1
    # via -r requirements.in
django-ckeditor==6.7.1
    # via -r requirements.in
django-colorfield==0.4.5
    # via -r requirements.in
django-cors-headers==3.10.0
    # via -r requirements.in
django-import-export==2.6.1
    # via -r requirements.in
django-js-asset==2.2.0
    # via django-ckeditor
django-material==1.10.0
    # via -r requirements.in
django-simple-history==3.3.0
    # via -r requirements.in
django-timescaledb==0.2.12
    # via -r requirements.in
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
filelock==3.18.0
    # via pygeoapi
fiona==1.10.1
    # via geopandas
flask==3.1.1
    # via pygeoapi
geopandas==0.10.2
    # via -r requirements.in
greenlet==3.2.3
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via
    #   fiona
    #   flask
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   pygeoapi
jsonschema==4.17.3
    # via pygeoapi
kiwisolver==1.4.7
    # via matplotlib
kombu==5.2.2
    # via celery
lark==1.2.2
    # via pygeofilter
lxml==5.4.0
    # via pyoscar
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.4.3
    # via
    #   -r requirements.in
    #   cartopy
    #   metpy
metpy==1.3.1
    # via -r requirements.in
minio==7.2.15
    # via -r requirements.in
numpy==1.21.2
    # via
    #   -r requirements.in
    #   cartopy
    #   matplotlib
    #   metpy
    #   opencv-python
    #   pandas
    #   rasterio
    #   scipy
    #   shapely
    #   snuggs
    #   xarray
odfpy==1.4.1
    # via tablib
opencv-python==4.5.4.60
    # via -r requirements.in
openpyxl==3.1.5
    # via
    #   -r requirements.in
    #   tablib
packaging==25.0
    # via
    #   pint
    #   pooch
    #   xarray
pandas==2.0.0
    # via
    #   -r requirements.in
    #   geopandas
    #   metpy
    #   xarray
pillow==11.2.1
    # via matplotlib
pint==0.18
    # via
    #   -r requirements.in
    #   metpy
platformdirs==4.3.8
    # via pooch
pooch==1.8.2
    # via metpy
prompt-toolkit==3.0.51
    # via click-repl
psycopg2-binary==2.9.2
    # via -r requirements.in
pycparser==2.22
    # via cffi
pycryptodome==3.23.0
    # via minio
pydantic==1.10.22
    # via pygeoapi
pygeoapi==0.17.0
    # via -r requirements.in
pygeofilter==0.3.1
    # via pygeoapi
pygeoif==1.5.1
    # via
    #   pygeoapi
    #   pygeofilter
pyjwt==2.4.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyoscar==0.9.0
    # via -r requirements.in
pyparsing==3.2.3
    # via
    #   matplotlib
    #   snuggs
pyproj==3.2.1
    # via
    #   -r requirements.in
    #   cartopy
    #   geopandas
    #   metpy
    #   pygeoapi
pyrsistent==0.20.0
    # via jsonschema
pyshp==2.3.1
    # via cartopy
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   dateparser
    #   matplotlib
    #   pandas
    #   pygeoapi
    #   python-crontab
python-memcached==1.59
    # via -r requirements.in
python-slugify==5.0.2
    # via -r requirements.in
pytz==2022.7
    # via
    #   -r requirements.in
    #   celery
    #   dateparser
    #   django
    #   django-timezone-field
    #   pandas
    #   pygeoapi
pyyaml==6.0.2
    # via
    #   pygeoapi
    #   tablib
rasterio==1.2.10
    # via
    #   -r requirements.in
    #   pygeoapi
redis==3.5.3
    # via -r requirements.in
regex==2024.11.6
    # via dateparser
requests==2.26.0
    # via
    #   -r requirements.in
    #   pooch
    #   pygeoapi
    #   pyoscar
scipy==1.10.1
    # via metpy
shapely==2.0.7
    # via
    #   cartopy
    #   geopandas
    #   pygeoapi
six==1.17.0
    # via
    #   django-material
    #   python-dateutil
    #   python-memcached
snuggs==1.4.7
    # via rasterio
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==1.4.54
    # via pygeoapi
sqlparse==0.5.3
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.8.0
    # via django-import-export
text-unidecode==1.3
    # via python-slugify
tinydb==4.8.2
    # via pygeoapi
traitlets==5.14.3
    # via metpy
typing-extensions==4.14.0
    # via
    #   asgiref
    #   beautifulsoup4
    #   minio
    #   pydantic
    #   pygeoif
tzlocal==5.3.1
    # via dateparser
unicodecsv==0.14.1
    # via pygeoapi
urllib3==1.26.20
    # via
    #   -r requirements.in
    #   minio
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via flask
xarray==2023.1.0
    # via metpy
xlrd==2.0.2
    # via tablib
xlwt==1.3.0
    # via tablib
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools


opencv-python==4.5.4.58

aquacrop==3.0.11
supabase==2.22.0
